Remove all the code for showing the "comment add" component in the feed.
Refactored the code a little so that the comment widget lives outside of the event item controls.
