--- src/app/components/event-item/controls/comments/add-placeholder/add-placeholder.vue
@@ -1,81 +1,0 @@
-<script lang="ts" src="./add-placeholder"></script>
-
-<template>
-	<div class="form-control -container" @click="onClick('')">
-		<div class="-placeholder">
-			<em>Leave a comment...</em>
-			<div class="-controls" @click.stop>
-				<app-jolticon
-					v-app-tooltip="$gettext('Insert GIF')"
-					class="-icon-btn"
-					icon="gif"
-					@click.native="onClick('gif')"
-				/>
-				<span
-					v-app-tooltip="$gettext('Insert Emoji')"
-					:class="'emoji-button emoji emoji-' + emoji"
-					@mouseenter="onMouseEnterEmoji"
-					@click="onClick('emoji')"
-				/>
-			</div>
-		</div>
-	</div>
-</template>
-
-<style lang="stylus" scoped>
-@import '~styles/variables'
-@import '~styles-lib/mixins'
-
-.-container
-	height: auto
-	cursor: text
-	position: relative
-	theme-prop('color', 'fg-muted')
-	change-bg('bg-offset')
-
-.-placeholder
-	padding-left: 4px
-	white-space: nowrap
-	padding-bottom: ($grid-gutter-width / 10)
-	padding-top: ($grid-gutter-width / 10)
-	display: flex
-	align-items: center
-	justify-content: space-between
-	position: relative
-
-.-controls
-	position: relative
-	max-height: $line-height-computed
-	display: flex
-	align-items: center
-
-.-icon-btn
-	font-size: 20px
-	margin-right: 12px
-	transition: filter 0.15s, transform 0.2s ease
-	cursor: pointer
-	color: var(--theme-fg)
-
-	&:not(:hover)
-		filter: grayscale(1) opacity(75%)
-
-	&:hover
-		transform: scale(1.2)
-
-	&:focus
-		outline: none
-
-.emoji-button
-	transition: filter 0.15s, transform 0.2s ease
-	cursor: pointer
-	image-rendering: pixelated
-
-	&:not(:hover)
-		filter: grayscale(1) opacity(75%)
-
-	&:hover
-		transform: scale(1.2)
-
-	&:focus
-		outline: none
-</style>
